+++ examples/convert_eeg_to_bids.py
@@ -47,9 +47,6 @@ from mne_bids.utils import print_dir_tree
 # Conveniently, there is already a data loading function available with
 # MNE-Python:
 
-# get MNE directory w/ example data
-mne_dir = mne.get_config('MNE_DATASETS_SAMPLE_PATH')
-
 # Define which tasks we want to download.
 tasks = [2,  # This is 2 minutes of eyes closed rest
          4,  # This is run #1 of imagining to close left or right fist
@@ -59,14 +56,16 @@ tasks = [2,  # This is 2 minutes of eyes closed rest
 for subj_idx in [1, 2]:
     eegbci.load_data(subject=subj_idx,
                      runs=tasks,
-                     path=mne_dir,
                      update_path=True)
 
 ###############################################################################
 # Let's see whether the data has been downloaded using a quick visualization
 # of the directory tree.
 
-data_dir = os.path.join(mne_dir, 'MNE-eegbci-data')
+# get MNE directory w/ example data
+mne_data_dir = mne.get_config('MNE_DATASETS_EEGBCI_PATH')
+
+data_dir = os.path.join(mne_data_dir, 'MNE-eegbci-data')
 print_dir_tree(data_dir)
 
 ###############################################################################
@@ -117,7 +116,7 @@ print(write_raw_bids.__doc__)
 subject_id = '001'  # zero padding to account for >100 subjects in this dataset
 task = 'resteyesclosed'
 raw_file = raw
-bids_root = os.path.join(mne_dir, 'eegmmidb_bids')
+bids_root = os.path.join(mne_data_dir, 'eegmmidb_bids')
 
 ###############################################################################
 # Now we just need to specify a few more EEG details to get something sensible: